requirements.txt
streamlit>=1.12.0
opencv-python-headless>=4.5.0
numpy>=1.19.0
Pillow>=9.0.0
streamlit-webrtc>=0.47.0
av
